--- src/test/java/ultimate_to_do/List.java
@@ -1,4 +1,0 @@
-package ultimate_to_do;
-
-public class List {
-}
