+++ src/test/java/ultimate_to_do/ListManager.java
@@ -1,0 +1,25 @@
+package ultimate_to_do;
+
+public class ListManager {
+
+    //constructor
+
+    //parameterized constructor
+
+    //deep copy constructor
+
+    //GETTERS
+
+    //getList()
+
+    //add todoItem
+
+    //remove todoItem
+
+    //clear List
+
+    //save list
+
+    //deep deconstruct List
+
+}
