+++ src/test/java/ultimate_to_do/PriorityListManager.java
@@ -1,0 +1,16 @@
+package ultimate_to_do;
+
+//PriorityListManager has a priorityList vector
+//
+public class PriorityListManager extends ListManager{
+
+    //priorityList is a vector that holds todoItems
+
+
+    //PRIVATE HELPER FUNCTIONS
+
+    //sortOnPriorityValue()
+
+
+
+}
